--- cl-log-api/src/main/kotlin/org/depromeet/cllog/server/api/security/SecurityConfig.kt
@@ -1,28 +1,0 @@
-package org.depromeet.cllog.server.api.security
-
-import org.depromeet.cllog.server.api.security.jwt.JwtFilter
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-import org.springframework.security.config.annotation.web.builders.HttpSecurity
-import org.springframework.security.config.http.SessionCreationPolicy
-import org.springframework.security.web.SecurityFilterChain
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter
-
-@Configuration
-class SecurityConfig(
-    private val jwtFilter: JwtFilter
-) {
-    @Bean
-    fun securityFilterChain(http: HttpSecurity): SecurityFilterChain {
-        http
-            .csrf { it.disable() }
-            .sessionManagement { it.sessionCreationPolicy(SessionCreationPolicy.STATELESS) }
-            .authorizeHttpRequests { auth ->
-                auth.requestMatchers("/api/v1/auth/**").permitAll()
-                auth.anyRequest().authenticated()
-            }
-            .addFilterBefore(jwtFilter, UsernamePasswordAuthenticationFilter::class.java)
-
-        return http.build()
-    }
-}
