--- cl-log-domain/src/main/kotlin/org/depromeet/cllog/server/domain/auth/application/dto/AppleUserInfo.kt
@@ -1,6 +1,0 @@
-package org.depromeet.cllog.server.domain.auth.application.dto
-
-data class AppleUserInfo(
-    val id: String,
-    val name: String
-)
