--- cl-log-domain/src/main/kotlin/org/depromeet/cllog/server/domain/auth/application/dto/LoginDetails.kt
@@ -1,8 +1,0 @@
-package org.depromeet.cllog.server.domain.auth.application.dto
-
-import org.depromeet.cllog.server.domain.user.domain.Provider
-
-data class LoginDetails(
-    val loginId: String,
-    val provider: Provider
-)
