--- cl-log-domain/src/main/kotlin/org/depromeet/cllog/server/domain/auth/presentation/exception/AuthException.kt
@@ -1,8 +1,0 @@
-package org.depromeet.cllog.server.domain.auth.presentation.exception
-
-import org.depromeet.cllog.server.domain.common.ClogException
-
-class AuthException(
-    errorCode: AuthErrorCode,
-    cause: Throwable? = null
-) : ClogException(errorCode, cause.toString())
