--- cl-log-domain/src/main/kotlin/org/depromeet/cllog/server/domain/user/domain/Provider.kt
@@ -1,5 +1,0 @@
-package org.depromeet.cllog.server.domain.user.domain
-
-enum class Provider {
-    KAKAO, APPLE
-}
