--- cl-log-domain/src/main/kotlin/org/depromeet/cllog/server/domain/user/domain/User.kt
@@ -1,23 +1,0 @@
-package org.depromeet.cllog.server.domain.user.domain
-
-import jakarta.persistence.*
-
-@Entity
-@Table(name = "users")
-class User(
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    val id: Long = 0,
-
-    @Column(nullable = false, unique = true)
-    val loginId: String,
-
-    val name: String,
-
-    @Enumerated(EnumType.STRING)
-    @Column(nullable = false)
-    val provider: Provider,
-
-    @Column(nullable = false)
-    var isDeleted: Boolean = false
-)
