+++ clog-api/src/main/kotlin/org/depromeet/clog/server/api/configuration/ClogControllerAdvice.kt
@@ -10,7 +10,7 @@ import org.springframework.web.bind.annotation.RestControllerAdvice
 class ClogControllerAdvice {
 
     @ExceptionHandler(ClogException::class)
-    fun handleClLogException(e: ClogException): ResponseEntity<ApiResponse<Unit>> {
+    fun handleclogException(e: ClogException): ResponseEntity<ApiResponse<Unit>> {
         return ResponseEntity.status(e.errorCode.httpStatus)
             .body(ApiResponse.fail(e))
     }
