+++ clog-domain/src/main/kotlin/org/depromeet/clog/server/domain/auth/application/dto/AppleUserInfo.kt
@@ -1,0 +1,6 @@
+package org.depromeet.clog.server.domain.auth.application.dto
+
+data class AppleUserInfo(
+    val id: String,
+    val name: String
+)
