+++ clog-domain/src/main/kotlin/org/depromeet/clog/server/domain/auth/application/dto/LoginDetails.kt
@@ -1,0 +1,8 @@
+package org.depromeet.clog.server.domain.auth.application.dto
+
+import org.depromeet.clog.server.domain.user.domain.Provider
+
+data class LoginDetails(
+    val loginId: String,
+    val provider: Provider
+)
