+++ clog-domain/src/main/kotlin/org/depromeet/clog/server/domain/auth/infrastructure/RefreshTokenRepository.kt
@@ -1,0 +1,9 @@
+package org.depromeet.clog.server.domain.auth.infrastructure
+
+import org.depromeet.clog.server.domain.auth.domain.RefreshToken
+
+interface RefreshTokenRepository {
+    fun save(token: RefreshToken): RefreshToken
+    fun findById(id: String): RefreshToken?
+    fun delete(token: RefreshToken)
+}
