+++ clog-domain/src/main/kotlin/org/depromeet/clog/server/domain/auth/presentation/exception/AuthException.kt
@@ -1,0 +1,8 @@
+package org.depromeet.clog.server.domain.auth.presentation.exception
+
+import org.depromeet.clog.server.domain.common.ClogException
+
+class AuthException(
+    errorCode: AuthErrorCode,
+    cause: Throwable? = null
+) : ClogException(errorCode, cause.toString())
