+++ clog-domain/src/main/kotlin/org/depromeet/clog/server/domain/user/domain/Provider.kt
@@ -1,0 +1,5 @@
+package org.depromeet.clog.server.domain.user.domain
+
+enum class Provider {
+    KAKAO, APPLE
+}
