+++ clog-domain/src/main/kotlin/org/depromeet/clog/server/domain/user/domain/User.kt
@@ -1,0 +1,23 @@
+package org.depromeet.clog.server.domain.user.domain
+
+import jakarta.persistence.*
+
+@Entity
+@Table(name = "users")
+class User(
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    val id: Long = 0,
+
+    @Column(nullable = false, unique = true)
+    val loginId: String,
+
+    val name: String,
+
+    @Enumerated(EnumType.STRING)
+    @Column(nullable = false)
+    val provider: Provider,
+
+    @Column(nullable = false)
+    var isDeleted: Boolean = false
+)
