+++ clog-global-utils/src/main/kotlin/org/depromeet/clog/server/global/utils/config/RestTemplateConfig.kt
@@ -1,0 +1,13 @@
+package org.depromeet.clog.server.global.utils.config
+
+import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Configuration
+import org.springframework.web.client.RestTemplate
+
+@Configuration
+class RestTemplateConfig {
+    @Bean
+    fun restTemplate(): RestTemplate {
+        return RestTemplate()
+    }
+}
