+++ clog-infrastructure/build.gradle.kts
@@ -1,5 +1,6 @@
 dependencies {
     implementation(project(":clog-global-utils"))
+    implementation(project(":clog-domain"))
 
     implementation("org.springframework.boot:spring-boot-starter-web")
     implementation("org.springframework.boot:spring-boot-starter-data-jpa")
