+++ clog-infrastructure/src/main/kotlin/org/depromeet/clog/server/infrastructure/user/UserRepositoryAdapter.kt
@@ -1,0 +1,22 @@
+package org.depromeet.clog.server.infrastructure.user
+
+import org.depromeet.clog.server.domain.user.domain.Provider
+import org.depromeet.clog.server.domain.user.domain.User
+import org.depromeet.clog.server.domain.user.infrastructure.UserRepository
+import org.springframework.stereotype.Component
+import java.util.*
+
+@Component
+class UserRepositoryAdapter(
+    private val jpaUserRepository: JpaUserRepository
+) : UserRepository {
+
+    override fun findByLoginIdAndProvider(loginId: String, provider: Provider): Optional<User> =
+        jpaUserRepository.findByLoginIdAndProvider(loginId, provider)
+
+    override fun findByLoginId(loginId: String): Optional<User> =
+        jpaUserRepository.findByLoginId(loginId)
+
+    override fun save(user: User): User =
+        jpaUserRepository.save(user)
+}
