+++ config/detekt.yml
@@ -23,3 +23,13 @@ performance:
 style:
   MagicNumber:
     active: false
+  WildcardImport:
+    active: false
+
+exceptions:
+  TooGenericExceptionCaught:
+    active: false
+
+formatting:
+  NoWildcardImports:
+    active: false
